1) Репозиторий пользователей 2) Добавил сообщение о введении неправильного логина на страницу авторизаци
