+++ src/core/baseRepositories/baseUserRepository.js
@@ -1,0 +1,8 @@
+const baseUserRepository =
+class {
+	constructor(userStorage) {
+		this.userStorage = userStorage;
+	}
+};
+
+module.exports = baseUserRepository;
