+++ src/core/useCases/tryLogin.js
@@ -1,9 +1,11 @@
 const BaseUseCase = require('./baseUseCase');
+const UserRepository = require('../../infrastructure/repositories/userRepository');
+
 
 const TryLoginUseCase =
 class extends BaseUseCase{
 	static execute(params, req){
-		const success = params.userid === 'asdasd';
+		const success = UserRepository.checkUserExistByID(params.userid);
 		if (success){
 			req.session.isLogin = true;
 			req.session.user = params.userid;
