+++ src/public/css/login.css
@@ -4,3 +4,8 @@
     margin-right: auto;
     margin-left: auto;
 }
+
+#error{
+    color: red;
+    visibility: hidden;
+}
