+++ src/public/templates/login.html
@@ -17,6 +17,7 @@
                     <label for="exampleInputEmail1" class="form-label">Введите свой логин:</label>
                     <input type="text" class="form-control" id="exampleInputEmail1" aria-describedby="emailHelp" name="userid">
                     <div id="emailHelp" class="form-text">Введите свой логин</div>
+                    <div id="error">Введён неверный логин</div>
                 </div>
                 <button type="submit" class="btn btn-primary">Войти</button>
             </form>
